fetch show time adapter + tests

+++ src/main/kotlin/com/mziolo/cinema/infrastructure/catalog/adapter/FetchMovieAdapter.kt
@@ -40,5 +40,5 @@ private fun ImdbMovieDto.toMovie(movieId: MovieId): Movie = Movie(
         .find { it.source == imdbSource }
         .let { valueRegex.find(it!!.value) }
         .let { it!!.groupValues[1].toDouble()},
-    runtimeInMinutes = runtime.removeSuffix(minSuffix).toLong()
+    runtimeInMinutes = runtime.removeSuffix(minSuffix).toInt()
 )
